Изменил в регистрах, надо бы запилить в модбасе.

+++ User/data_model.h
@@ -82,13 +82,13 @@
 #define H_SENSOR_TYPE         ( CO2_SENSOR_TYPE + sizeof(uint8_t) )
 #define F_CHANNEL             ( H_SENSOR_TYPE   + sizeof(uint8_t) )
 #define SETTING_MAX           ( F_CHANNEL       + sizeof(uint32_t) )
-#define SETTING_MIN           ( SETTING_MAX       + sizeof(uint32_t) )
-#define SETTING_MID           ( SETTING_MIN       + sizeof(uint32_t) )
-#define INPUT_SENSOR_TYPE     ( SETTING_MID       + sizeof(uint32_t) )
+#define SETTING_MIN           ( SETTING_MAX       + sizeof(uint16_t) )
+#define SETTING_MID           ( SETTING_MIN       + sizeof(uint16_t) )
+#define INPUT_SENSOR_TYPE     ( SETTING_MID       + sizeof(uint16_t) )
 #define OFFSET_CH2            ( INPUT_SENSOR_TYPE       + sizeof(uint8_t) )
-#define CH1_SETTING           ( OFFSET_CH2         + sizeof(uint32_t) )
-#define CH2_SETTING           ( CH1_SETTING         + sizeof(uint32_t) )
-#define FILTER_LOW            ( CH2_SETTING         + sizeof(uint32_t) )
+#define CH1_SETTING           ( OFFSET_CH2         + sizeof(uint16_t) )
+#define CH2_SETTING           ( CH1_SETTING         + sizeof(uint16_t) )
+#define FILTER_LOW            ( CH2_SETTING         + sizeof(uint16_t) )
 #define FILTER_HIGH           ( FILTER_LOW + sizeof(uint16_t) )
 #define MOTO_HOURS            ( FILTER_HIGH + sizeof(uint16_t) )
 #define RECORD_COUNT          ( MOTO_HOURS  + sizeof(uint32_t)  )
@@ -98,6 +98,10 @@
 #define LIGTH                 ( SYSTEM_START + sizeof(uint8_t) )
 #define MODE                  ( LIGTH    + sizeof(uint8_t) )
 #define SENSOR_ERROR          ( MODE  + sizeof(uint8_t) )
+
+
+
+
 
 #define RECORD_SIZE 7
 #define RECORD_DATA_SIZE      50
